fix: update currentPos when seekTo called
The current position is tracked separately from the player and requires
the player to stream an updated position. However, if the player is
paused and seekTo is called, then the currentPos is out of sync with the
players position.

This change explicitly updates the current position being tracked so
that it can be used until the player starts again and updates the
current position.

Signed-off-by: Nick Campbell <nicholas.j.campbell@gmail.com>